+++ eq-author/src/components/ImportContentQuestionnairesView/QuestionnairesTable/index.js
@@ -6,10 +6,10 @@ import styled from "styled-components";
 
 import { SORT_ORDER } from "constants/sort-order.js";
 
-import QuestionnaireTable from "components/QuestionnaireTable";
+import QuestionnaireTable from "components/ImportContentQuestionnaireTable";
 import Panel from "components/Panel";
 
-import tableHeadings from "components/QuestionnaireTable/TableHeadings";
+import tableHeadings from "components/ImportContentQuestionnaireTable/TableHeadings";
 
 import { useQuestionnaireLockingModal } from "components/modals/QuestionnaireLockingModal";
