+++ eq-author/src/components/modals/QuestionnaireSelectModal/index.js
@@ -9,7 +9,7 @@ import { colors } from "constants/theme";
 import Modal from "components/modals/Modal";
 import Button from "components/buttons/Button";
 import ButtonGroup from "components/buttons/ButtonGroup";
-import QuestionnairesView from "components/QuestionnairesView";
+import QuestionnairesView from "components/ImportContentQuestionnairesView";
 
 const ModalFooter = styled.footer`
   padding: 1.5em;
